backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
motor==3.1.2
pydantic==1.10.7
python-dotenv==1.0.0
pymongo==4.3.3
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.0.0
